backend/requirements.txt
langchain==0.3.7
langchain-core==0.3.15
langchain-groq==0.2.1
langgraph==0.2.50
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
pandas==2.2.3
plotly==5.22.0
scipy==1.13.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10